backend/pyproject.toml
[tool.poetry]
name = "career-assistant-backend"
version = "1.0.0"
description = "Backend API for Career Assistant application"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "lib"}, {include = "*.py"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.104.1"
uvicorn = {extras = ["standard"], version = "0.24.0"}
python-multipart = "0.0.6"
supabase = "2.0.0"
google-generativeai = ">=0.8.0"
reportlab = "4.0.7"
python-dotenv = "1.0.0"
httpx = ">=0.24.0,<0.25.0"
pydantic = "2.5.0"
aiofiles = "23.2.1"
pypdf2 = "^3.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.0"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
# Select specific rules to enable/disable if needed

